+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 ### Changed
 - Updates README for usage with Rails
 
+### Fixed
+- Handles parsing exceptions gracefully
+
 ## [0.2.0] - 2016-12-13
 ### Added
 - Occurrence exposes content type